# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Travel Planner
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type TravelPreferencesBudget = typeof TravelPreferencesBudget[keyof typeof TravelPreferencesBudget];


export const TravelPreferencesBudget = {
  budget: 'budget',
  'mid-range': 'mid-range',
  luxury: 'luxury',
} as const;

export type TravelPreferencesTravelStyle = typeof TravelPreferencesTravelStyle[keyof typeof TravelPreferencesTravelStyle];


export const TravelPreferencesTravelStyle = {
  adventure: 'adventure',
  cultural: 'cultural',
  relaxation: 'relaxation',
  foodie: 'foodie',
  family: 'family',
  romantic: 'romantic',
  backpacker: 'backpacker',
} as const;

export interface TravelPreferences {
  /** Destination city/country */
  destination: string;
  /** Number of days for the trip */
  durationDays: number;
  budget: TravelPreferencesBudget;
  /** Number of travelers */
  travelers: number;
  travelStyle: TravelPreferencesTravelStyle;
  /**
     * Optional start date (YYYY-MM-DD)
     * @nullable
     */
  startDate?: string | null;
  /** Specific interests or activities */
  interests?: string[];
  /**
     * Any special requirements (dietary, accessibility, etc.)
     * @nullable
     */
  specialRequirements?: string | null;
  /**
     * Origin city/location for intercity transport planning
     * @nullable
     */
  transportFrom?: string | null;
  /**
     * Destination city for intercity transport (defaults to destination if omitted)
     * @nullable
     */
  transportTo?: string | null;
}

export interface Itinerary {
  id: number;
  destination: string;
  durationDays: number;
  budget: string;
  travelers: number;
  travelStyle: string;
  /** @nullable */
  startDate?: string | null;
  /**
     * JSON-serialized array of interests
     * @nullable
     */
  interests?: string | null;
  /** @nullable */
  specialRequirements?: string | null;
  /** @nullable */
  transportFrom?: string | null;
  /** @nullable */
  transportTo?: string | null;
  /**
     * Destination hero image URL (from Wikipedia)
     * @nullable
     */
  imageUrl?: string | null;
  /** Full AI-generated itinerary text (markdown) */
  content: string;
  createdAt: string;
}

export interface ItineraryInput {
  destination: string;
  durationDays: number;
  budget: string;
  travelers: number;
  travelStyle: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  interests?: string | null;
  /** @nullable */
  specialRequirements?: string | null;
  /** @nullable */
  transportFrom?: string | null;
  /** @nullable */
  transportTo?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  content: string;
}

export type ItineraryStatsTopDestinationsItem = {
  destination: string;
  count: number;
};

export type ItineraryStatsBudgetBreakdown = {
  budget: number;
  midRange: number;
  luxury: number;
};

export type ItineraryStatsPopularStylesItem = {
  style: string;
  count: number;
};

export interface ItineraryStats {
  totalItineraries: number;
  topDestinations: ItineraryStatsTopDestinationsItem[];
  budgetBreakdown: ItineraryStatsBudgetBreakdown;
  popularStyles: ItineraryStatsPopularStylesItem[];
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface OpenaiConversationInput {
  title: string;
}

export interface OpenaiMessageInput {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export interface OpenaiError {
  error: string;
}