# lib/api-zod/src/generated/types/itinerary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Travel Planner
 * OpenAPI spec version: 0.1.0
 */

export interface Itinerary {
  id: number;
  destination: string;
  durationDays: number;
  budget: string;
  travelers: number;
  travelStyle: string;
  /** @nullable */
  startDate?: string | null;
  /**
     * JSON-serialized array of interests
     * @nullable
     */
  interests?: string | null;
  /** @nullable */
  specialRequirements?: string | null;
  /** @nullable */
  transportFrom?: string | null;
  /** @nullable */
  transportTo?: string | null;
  /**
     * Destination hero image URL (from Wikipedia)
     * @nullable
     */
  imageUrl?: string | null;
  /** Full AI-generated itinerary text (markdown) */
  content: string;
  createdAt: Date;
}