# lib/api-zod/src/generated/types/itineraryInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification for AI Travel Planner
 * OpenAPI spec version: 0.1.0
 */

export interface ItineraryInput {
  destination: string;
  durationDays: number;
  budget: string;
  travelers: number;
  travelStyle: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  interests?: string | null;
  /** @nullable */
  specialRequirements?: string | null;
  /** @nullable */
  transportFrom?: string | null;
  /** @nullable */
  transportTo?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  content: string;
}